Transform Overlay Header to a component
