+++ frontend/src/Pages/DishOverviewPage.js
@@ -1,8 +1,9 @@
+import PropTypes from 'prop-types'
+import { FaCheck } from 'react-icons/fa'
 import { Link } from 'react-router-dom'
-import { FaChevronRight, FaCheck } from 'react-icons/fa'
 import styled from 'styled-components/macro'
 import DishItem from '../Components/DishItem'
-import PropTypes from 'prop-types'
+import HeaderOverlay from '../Components/HeaderOverlay'
 
 DishOverview.propTypes = {
   dishes: PropTypes.array,
@@ -16,10 +17,9 @@ export default function DishOverview({ dishes, onToggleItem }) {
 
   return (
     <Wrapper>
-      <Header>
-        <LinkStyled to="/home">X</LinkStyled>
+      <HeaderOverlay>
         <p>Dish Overview</p>
-      </Header>
+      </HeaderOverlay>
       <Grid>
         {dishes &&
           dishes.map(({ id, name, isSelected }) => {
@@ -60,27 +60,6 @@ export default function DishOverview({ dishes, onToggleItem }) {
 const Wrapper = styled.div`
   display: grid;
   grid-template-rows: 60px auto;
-`
-//TODO make the Header a Component
-
-const Header = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-evenly;
-  z-index: 10;
-  background-color: var(--c-green);
-  box-shadow: 0px 0px 6px 3px rgba(0, 0, 0, 0.08);
-  border-radius: 0 0 15px 15px;
-  position: fixed;
-  top: 0;
-  width: 100%;
-`
-const LinkStyled = styled(Link)`
-  text-decoration: none;
-  color: var(--c-gray);
-  font-size: 1rem;
-  position: fixed;
-  left: 20px;
 `
 
 const Grid = styled.div`
